requirements.txt
duckdb==1.5.2
pyarrow==24.0.0
dbt-duckdb==1.9.0 
pandas==2.2.3 
numpy==2.1.3 
rapidfuzz==3.6.1 
requests==2.31.0 
beautifulsoup4==4.12.3 
azure-storage-file-datalake==12.14.0 
azure-identity==1.15.0 
python-dotenv==1.0.0 
adlfs==2024.2.0 
